vyatta_swport: pass double-pointer to plugin_tx_framer plugin API
The plugin may change the mbuf pointer, such as if required to add a
large amount of encap onto the mbuf before sending into the backplane
port, or because a function that is called doesn't have exactly the
same checks for a shared mbuf as in the prep_header_change function.

Therefore, change the plugin_tx_framer API such that it takes a
double-pointer to the mbuf so we can upate the one in the array that
will be transmitted later on switch_port_tx.

Fixes: VRVDR-49246