Add API to retrieve information about the sw_port device
Add an API to retrieve a FAL object handle for it, if applicable, and
whether or not it is a LAG bundle. These are necessary to communicate
association between FAL API calls and the PMD without encoding this
information in the FAL API, which need not and should not be dependent
on DPDK port representations of interfaces.

VRVDR-49262