Makefile: don't use -march=native
The machine that the package is being built on may be newer and
support more instructions than the device the code is eventually run
on. Therefore, don't use the -march=native compiler flag - the
architecture to be optimised for will instead by inherited from the
DPDK CFLAGS.

In addition, pass in -O3 since there is code here which is in a hot
path for certain use cases.

Fixes: VRVDR-49400